// src/shared/model/shared-user.model.ts
import { z } from 'zod';
import { UserStatus } from '../constants/auth.constant';

export const UserSchema = z.object({
  id: z.number(),
  email: z
    .string({ message: 'Error.InvalidEmail' })
    .email({ message: 'Error.InvalidEmail' }),
  name: z.string().min(1, 'Error.InvalidName').max(100, 'Error.InvalidName'),
  password: z.string().min(6).max(100),
  phoneNumber: z.string().min(9).max(15),
  avatar: z.string().nullable(),
  totpSecret: z.string().nullable(),
  status: z.enum([UserStatus.ACTIVE, UserStatus.INACTIVE, UserStatus.BLOCKED]),
  roleId: z.number().positive(),
  createdById: z.number().nullable(),
  updatedById: z.number().nullable(),
  deletedById: z.number().nullable(),
  deletedAt: z.date().nullable(),
  createdAt: z.date(),
  updatedAt: z.date(),
});

// export const GetUserProfileResSchema = UserSchema.omit({
//   password: true,
//   totpSecret: true,
// }).extend({
//   role: RoleSchema.pick({
//     id: true,
//     name: true,
//   }).extend({
//     permissions: z.array(
//       PermissionSchema.pick({
//         id: true,
//         name: true,
//         module: true,
//         path: true,
//         method: true,
//       }),
//     ),
//   }),
// });

// export const UpdateProfileResSchema = UserSchema.omit({
//   password: true,
//   totpSecret: true,
// });

export type UserType = z.infer<typeof UserSchema>;
// export type GetUserProfileResType = z.infer<typeof GetUserProfileResSchema>;
// export type UpdateProfileResType = z.infer<typeof UpdateProfileResSchema>;